requirements.txt
mcp>=1.2.0
httpx>=0.24.0
flask[async]>=3.1.0
python-dotenv>=0.19.0 